Clock setup self wire diagnostic, debug setup mode

+++ src/src/libraries/DS3231/DS3231.h
@@ -152,7 +152,7 @@ class DS3231
 	uint8_t conv2d(const char* p);
 
 	void writeRegister8(uint8_t reg, uint8_t value);
-	uint8_t readRegister8(uint8_t reg);
+	int readRegister8(uint8_t reg);
 };
 
 #endif
